requirements.txt
datasets
pandas
numpy
sentence-transformers
openai
python-dotenv
torch